arm64: dts: google: Convert DT overlays to full DTBs
DT overlays are a hassle to deal with when it comes to building and
especially shipping. Part of the kernel residing on a separate partition
is harder to manage, and it breaks convenient development commands such
as "fastboot boot" and "fastboot flash boot". This is especially an
issue for custom kernels, since users can end up in a broken state
where the boot and dtbo images are incompatible.

Qualcomm's LinuxLoader BootLib will actually use an appended DTB and
ignore DTBOs if it has the correct board ID, so we can leverage that to
use full DTBs instead.

Only production device revisions are supported.

DTBOs can still be built if desired with the CONFIG_BUILD_ARM64_DT_OVERLAY
config option.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>