google_battery: Fix SSOC delta's misuse of qnum API
qnum_rconst is only meant to be used with compile-time constants because
it relies on compile-time floating-point arithmetic that cannot be done
at runtime in the kernel. qnum_fromint is to be used for non-constant
values instead, but it gives a slightly different result due to not
adding 0.5.

Remove the ability to change ssoc_delta in DT and hard-code SSOC_DELTA
to the value specified in redbull's DT to fix the qnum API misuse.

This fixes the following error when compiling with GCC:

../drivers/power/supply/google/google_battery.c: In function 'ssoc_change_curve':
../drivers/power/supply/google/google_battery.c:744:6: error: '-mgeneral-regs-only' is incompatible with the use of flo
ating-point types
  744 | void ssoc_change_curve(struct batt_ssoc_state *ssoc_state,
      |      ^~~~~~~~~~~~~~~~~

Test: verify value of qnum_rconst(SSOC_DELTA)
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>